Fix type of recently introduced semi arbitrated sampling rate
Sammpling rates are not necessarily integers. Sure, they are in
the case of the MPU-6050, which is fortunately the only chip we
support at the moment with several interrupt driven sensors on
the same iio device, so this is not an issue for now, but still.

+++ common.h
@@ -375,7 +375,7 @@ typedef struct
 	 */
 	int needs_enable;
 
-	int semi_arbitrated_rate;	/* Arbitrated sampling rate before we considered other sensors co-located on the same iio device */
+	float semi_arbitrated_rate;	/* Arbitrated sampling rate before we considered other sensors co-located on the same iio device */
 }
 sensor_info_t;
